+++ python/cudf/cudf/core/frame.py
@@ -182,16 +182,29 @@ class Frame(libcudf.table.Table):
         # names with their integer positions in the `cols` list
         tables = []
         for cols in columns:
-            table_cols = cols[first_data_column_position:]
-            table_names = indices[first_data_column_position:]
-            table = cls(data=dict(zip(table_names, table_cols)))
+            table_index = None
             if 1 == first_data_column_position:
-                table._index = as_index(cols[0])
+                table_index = as_index(cols[0])
             elif first_data_column_position > 1:
-                index_cols = cols[:first_data_column_position]
-                index_names = indices[:first_data_column_position]
-                table._index = cls(data=dict(zip(index_names, index_cols)))
-            tables.append(table)
+                table_index = libcudf.table.Table(
+                    data=dict(
+                        zip(
+                            indices[:first_data_column_position],
+                            cols[:first_data_column_position],
+                        )
+                    )
+                )
+            tables.append(
+                libcudf.table.Table(
+                    data=dict(
+                        zip(
+                            indices[first_data_column_position:],
+                            cols[first_data_column_position:],
+                        )
+                    ),
+                    index=table_index,
+                )
+            )
 
         # Concatenate the Tables
         out = cls._from_table(
@@ -1178,6 +1191,13 @@ class Frame(libcudf.table.Table):
             To explicitly construct a GPU array, consider using \
             cupy.asarray(...)\nTo explicitly construct a \
             host array, consider using .to_array()"
+        )
+
+    def __arrow_array__(self, type=None):
+        raise TypeError(
+            "Implicit conversion to a host PyArrow Array via __arrow_array__ "
+            "is not allowed, To explicitly construct a PyArrow Array, "
+            "consider using .to_arrow()"
         )
 
     def drop_duplicates(