+++ python/cudf/cudf/utils/dtypes.py
@@ -223,7 +223,10 @@ def is_list_like(obj):
     Boolean: True or False depending on whether the
     input `obj` is like-like or not.
     """
-    return isinstance(obj, (Sequence,),) and not isinstance(obj, (str, bytes))
+
+    return isinstance(obj, (Sequence, np.ndarray)) and not isinstance(
+        obj, (str, bytes)
+    )
 
 
 def is_column_like(obj):
@@ -295,6 +298,19 @@ def min_signed_type(x, min_size=8):
                 return int_dtype
     # resort to using `int64` and let numpy raise appropriate exception:
     return np.int64(x).dtype
+
+
+def min_unsigned_type(x, min_size=8):
+    """
+    Return the smallest *unsigned* integer dtype
+    that can represent the integer ``x``
+    """
+    for int_dtype in np.sctypes["uint"]:
+        if (np.dtype(int_dtype).itemsize * 8) >= min_size:
+            if 0 <= x <= np.iinfo(int_dtype).max:
+                return int_dtype
+    # resort to using `uint64` and let numpy raise appropriate exception:
+    return np.uint64(x).dtype
 
 
 def min_column_type(x, expected_type):